requirements.txt
# LangChain 核心
langchain==0.2.10
langchain-openai==0.1.20
langchain-community==0.2.9

# 向量库
faiss-cpu==1.8.0

# 环境配置
python-dotenv==1.0.1

# Web 服务 & 接口
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.20.0

# 文件解析（PDF/Word/Excel/TXT）
pypdf==4.1.0
python-docx==1.1.0
pandas==2.2.1
openpyxl==3.1.2

# 网络 & 工具
requests
tavily-python

# 日志
loguru==0.7.2